Added a paillier library Implemented election board portion of signing stuff works with big ints now More validation

+++ src/BulletinBoard.java
@@ -1,3 +1,5 @@
+import java.math.BigInteger;
+
 /**
  * Created by james on 11/30/16.
  */
@@ -6,7 +8,7 @@ public class BulletinBoard {
 
   }
 
-  public void acceptAndZKPVote(long[] signedVote, Voter voter) throws BulletinBoardError{
+  public void acceptAndZKPVote(BigInteger[] signedVote, Voter voter) throws BulletinBoardError{
 
   }
 }
